# W4-library-version8/server/openapi.yaml
openapi: 3.0.3
info:
  title: Library Management REST API
  version: 3.0.0
  description: |
    REST API Server cho hệ thống quản lý thư viện.
    
    ## Đặc điểm
    - Kiến trúc Client-Server với REST principles
    - Stateless authentication với JWT token
    - Hỗ trợ HTTP caching với ETag và Cache-Control
    - CORS enabled cho phép client từ domain khác
    
    ## Nguyên tắc REST được áp dụng
    - ✅ **Stateless**: Không session, sử dụng JWT
    - ✅ **Client-Server**: Tách biệt hoàn toàn
    - ✅ **Uniform Interface**: HTTP methods chuẩn
    - ✅ **Resource-based**: URL định danh resources
    - ✅ **Cacheable**: HTTP caching headers
    - ✅ **JSON representation**: Dữ liệu trả về dạng JSON
    
    ## Authentication
    Hầu hết các endpoint yêu cầu authentication bằng JWT token.
    
    **Cách sử dụng:**
    1. Đăng nhập qua `/api/auth/login` để nhận token
    2. Thêm token vào header: `Authorization: Bearer <token>`
    3. Gửi request với token trong header
    
  contact:
    name: Library Management API Support
    email: support@library.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: http://127.0.0.1:5000
    description: Local server
...
paths:
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Đăng nhập và nhận JWT token
      description: |
        Xác thực người dùng và trả về JWT token.
        
        **REST Principles:**
        - Stateless: Token được tạo và trả về, server không lưu session
        - Client-Server: Server chỉ xử lý logic, client lưu token
        - Non-cacheable: POST request với sensitive data
        
        **Demo:** Hiện tại chấp nhận bất kỳ username/password nào.
        Trong production cần implement authentication thực.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: Tên đăng nhập
                  example: admin
                password:
                  type: string
                  format: password
                  description: Mật khẩu
                  example: admin123
            examples:
              admin:
                summary: Đăng nhập admin
                value:
                  username: admin
                  password: admin123
              user:
                summary: Đăng nhập user
                value:
                  username: user
                  password: password
      responses:
        '200':
          description: Đăng nhập thành công
          headers:
            Cache-Control:
              description: Không cache response
              schema:
                type: string
                example: no-store, no-cache, must-revalidate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT token để sử dụng cho các request tiếp theo
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      username:
                        type: string
                        example: admin
                      token_type:
                        type: string
                        example: Bearer
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/verify:
    get:
      tags:
        - Authentication
      summary: Kiểm tra token có hợp lệ không
      description: |
        Xác thực JWT token hiện tại.
        
        **REST Principles:**
        - Stateless: Verify từ token, không cần tra cứu session
        - Cacheable: Private cache 60s (chỉ browser)
        - Vary: Cache phụ thuộc vào Authorization header
      operationId: verifyToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token hợp lệ
          headers:
            Cache-Control:
              schema:
                type: string
                example: private, max-age=60
            Vary:
              schema:
                type: string
                example: Authorization
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      valid:
                        type: boolean
                        example: true
                      username:
                        type: string
                        example: admin
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/books:
    get:
      tags:
        - Books
      summary: Lấy danh sách sách
      description: |
        Lấy danh sách sách với pagination và filtering.
        
        **REST Principles:**
        - Stateless: Tất cả filtering qua query params
        - Uniform Interface: HTTP GET method
        - Self-descriptive: Response chứa metadata về pagination
        - Cacheable: Public cache 60s, ETag enabled
      operationId: getBooks
      parameters:
        - name: page
          in: query
          description: Số trang
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Số items mỗi trang (tối đa 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: author
          in: query
          description: Filter theo tác giả (partial match, không phân biệt hoa thường)
          schema:
            type: string
            example: Nguyen
        - name: title
          in: query
          description: Filter theo tiêu đề (partial match, không phân biệt hoa thường)
          schema:
            type: string
            example: Python
        - name: isbn
          in: query
          description: Filter theo ISBN (exact match)
          schema:
            type: string
            example: "9781234567890"
        - name: available_only
          in: query
          description: Chỉ lấy sách còn available
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Thành công
          headers:
            Cache-Control:
              schema:
                type: string
                example: public, max-age=60
            ETag:
              schema:
                type: string
                example: "33a64df551425fcc55e4d42a148795d9f25f89d4"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      books:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Books
      summary: Tạo sách mới
      description: |
        Tạo một cuốn sách mới trong thư viện.
        
        **Yêu cầu authentication.**
        
        **REST Principles:**
        - Stateless: Sử dụng JWT token, không session
        - Uniform Interface: POST để tạo resource mới
        - Non-cacheable: Mutation operation
      operationId: createBook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - isbn
                - quantity
              properties:
                title:
                  type: string
                  description: Tiêu đề sách
                  example: Clean Code
                author:
                  type: string
                  description: Tác giả
                  example: Robert C. Martin
                isbn:
                  type: string
                  description: Mã ISBN (phải unique)
                  example: "9780132350884"
                quantity:
                  type: integer
                  minimum: 1
                  description: Số lượng sách
                  example: 5
            examples:
              example1:
                summary: Sách lập trình
                value:
                  title: Clean Code
                  author: Robert C. Martin
                  isbn: "9780132350884"
                  quantity: 5
      responses:
        '201':
          description: Sách được tạo thành công
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store, no-cache, must-revalidate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Book created successfully
                  data:
                    $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/borrow-records/{record_id}:
    get:
      tags:
        - Borrow Records
      summary: Lấy thông tin chi tiết một bản ghi mượn sách
      description: |
        Lấy thông tin đầy đủ của một bản ghi mượn sách theo ID.
        
        **Yêu cầu authentication.**
        
        **REST Principles:**
        - Cacheable: Private cache 60s
      operationId: getBorrowRecord
      security:
        - bearerAuth: []
      parameters:
        - name: record_id
          in: path
          required: true
          description: ID của bản ghi mượn sách
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Thành công
          headers:
            Cache-Control:
              schema:
                type: string
                example: private, max-age=60
            Vary:
              schema:
                type: string
                example: Authorization
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/BorrowRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/borrow-records/{record_id}/return:
    put:
      tags:
        - Borrow Records
      summary: Trả sách
      description: |
        Đánh dấu sách đã được trả lại.
        
        **Yêu cầu authentication.**
        
        **REST Principles:**
        - Stateless: Chỉ cần record_id
        - Idempotent: Có thể gọi nhiều lần (lần 2 trở đi trả 409)
        - Atomic: Transaction đảm bảo consistency (tăng available count)
      operationId: returnBook
      security:
        - bearerAuth: []
      parameters:
        - name: record_id
          in: path
          required: true
          description: ID của bản ghi mượn sách
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Trả sách thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Book returned successfully
                  data:
                    $ref: '#/components/schemas/BorrowRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict - Sách đã được trả
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Conflict
                  message:
                    type: string
                    example: Book has already been returned
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication.
        
        **Cách lấy token:**
        1. Gọi POST `/api/auth/login` với username và password
        2. Lấy token từ response
        3. Thêm vào header: `Authorization: Bearer <token>`
        
        **Token expiration:** 24 giờ (mặc định)
...
  responses:
    BadRequest:
      description: Bad Request - Request không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_fields:
              summary: Thiếu trường bắt buộc
              value:
                error: Bad Request
                message: "Missing required fields: title, author"
            invalid_data:
              summary: Dữ liệu không hợp lệ
              value:
                error: Bad Request
                message: Quantity must be a positive integer
...